Auto-Length Adaptive Channels
Introduction

The key innovation of the ALAC is the implementation of dynamic length identification, which allows the indicator to adjust to the "market beat" or dominant cycle in real-time.
The Auto-Length Adaptive Channels (ALAC) is a flexible technical analysis tool that combines the benefits of five different approaches to market band and price deviation calculations.
Traders often tend to overthink of what length their indicators should use, and this is the main idea behind this script. It automatically calculates length based on pivot points, averaging the distance that is in between of current market highs and lows.
This approach is very helpful to identify market deviations, because deviations are always calculated and compared to previous market behavior.

How it works

The indicator uses a Detrended Rhythm Oscillator (DRO) to identify the dominant cycle in the market. This length information is then used to calculate different market bands and price deviations. The ALAC combines five different methodologies to compute these bands:

1 - Bollinger Bands
2 - Keltner Channels
3 - Envelope
4 - Average True Range Channels
5 - Donchian Channels

By averaging these calculations, the ALAC produces an overall market band that generalizes the approaches of these five methods into a single, adaptive channel.

How to Use

When the price is at the upper band, this might suggest that the asset is overbought and may be due for a price correction. Conversely, when the price is at the lower band, the asset may be oversold and due for a price increase.
The space between the bands represents the market's volatility. Wider bands indicate higher volatility, while narrower bands suggest lower volatility.
Indicator Settings

The settings of the ALAC allow for customization to suit different trading strategies:

Use Autolength?: This allows the indicator to automatically adjust the length of the dominant cycle.
Usual Length: If "Use Autolength?" is disabled, this setting allows the user to manually specify the length of the cycle.
Moving Average Type: This selects the type of moving average to be used in the calculations. Options include SMA, EMA, ALMA, DEMA, JMA, KAMA, SMMA, TMA, TSF, VMA, VAMA, VWMA, WMA, and ZLEMA.
Channel Multiplier: This adjusts the distance between the bands.
Channel Multiplier Step: This changes the step size of the channel multiplier. Each next market band will be multiplied by a previous one. You can potentially use values below 1, which will plot bands inside the first, main channel.
Use DPO instead of source data?: This setting uses the DPO for calculations instead of the source data. Basically, this is how you can add or eliminate trend from calculation of an average leg-up / leg-down move.
Fast: This adjusts the fast length of the DPO.
Slow: This adjusts the slow length of the DPO.
Zig-zag Period: This adjusts the period of the zig-zag pattern used in the DPO.

The Adaptive Average Channel is a powerful and versatile indicator that combines the strengths of multiple technical analysis methods.

In summary, with the ALAC, you can:

1 - Dynamically adapt to any asset and price action with automatic calculation of dominant cycle lengths.
2 - Identify potential overbought and oversold conditions with the adaptive market bands.
3 - Customize your analysis with various settings, including moving average type and channel multiplier.
4 - Enhance your trading strategy by using the indicator in conjunction with other forms of analysis.