Market Activity Risk
This indicator measures continuous momentum in miners market fee increase. 

This allows us to measure market activity - where people are majorly buying and selling. This allows us to point on levels of interest and align our decisions with overall market. 

Works on a daily timeframe, moving average is taken to compare deviations with previous 2 years of activity (assuming that cycles are taking about 4 years)